update script to autostart

+++ python_controller/shell/setup.sh
@@ -4,16 +4,18 @@ SCRIPT_DIR=$(cd $(dirname $0); pwd)
 echo '-----------install vsftpd-----------'
 sudo apt install vsftpd #FTPサーバのインストール
 FTP_CONFIG_TRGET=/etc/vsftpd.conf
-sudo cp $FTP_CONFIG_TRGET $FTP_CONFIG_TRGET'-default'
-TIME=$(date)
-echo "#-----------change by setup.bash: "$TIME"-----------" | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
-echo 'local_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
-echo 'write_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
-echo 'local_umask=022' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
-echo 'ascii_upload_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
-echo 'ascii_download_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
-cat $FTP_CONFIG_TRGET
-sudo service vsftpd start #FTPサーバの再起動
+if [ -e $FTP_CONFIG_TRGET];then
+	sudo cp $FTP_CONFIG_TRGET $FTP_CONFIG_TRGET'-default'
+	TIME=$(date)
+	echo "#-----------change by setup.bash: "$TIME"-----------" | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
+	echo 'local_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
+	echo 'write_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
+	echo 'local_umask=022' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
+	echo 'ascii_upload_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
+	echo 'ascii_download_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
+	cat $FTP_CONFIG_TRGET
+	sudo service vsftpd start #FTPサーバの再起動
+fi
 
 #####ショートカット用bashファイルの作成
 #echo '-----------install vsftpd-----------'
@@ -49,6 +51,7 @@ pip3 install watchdog
 
 ######set chmod of scripts/
 echo '-----------set chmod to scripts/-----------'
+sudo chmod 777 $SCRIPT_DIR/start_communication.sh
 sudo chmod -R 777 $SCRIPT_DIR/../scripts/
 
 ######setup UDP port
@@ -62,16 +65,18 @@ sudo ufw reload
 
 ######set autostart file
 echo '-----------setup auto start file-----------'
-cd ~/.config
+#cd ~/.config
 FTP_START_TARGET=ftp_start.desktop
-TARGET_SHELL=start_communication.bash
-mkdir autostart
-cd $_
-touch $FTP_START_TARGET
-#echo 'local_enable=YES' | sudo tee -a $FTP_CONFIG_TRGET>/dev/null
-echo 'Exec=lxterminal -e '$SCRIPT_DIR'/../shell/'$TARGET_SHELL | sudo tee -a $FTP_START_TARGET>/dev/null
-echo 'Type=Application' | sudo tee -a $FTP_START_TARGET>/dev/null
-echo 'Name=FTP_server' | sudo tee -a $FTP_START_TARGET>/dev/null
-echo 'Terminal=true' | sudo tee -a $FTP_START_TARGET>/dev/null
-
+TARGET_SHELL=start_communication.sh
+FTP_START_DIR=~/.config/autostart
+if [ ! -d $FTP_START_DIR ];then
+	mkdir $FTP_START_DIR 
+	cd $_
+	touch $FTP_START_TARGET
+	echo '[Desktop Entry]' | sudo tee -a $FTP_START_TARGET>/dev/null
+	echo 'Exec=lxterminal -e '$SCRIPT_DIR'/../shell/'$TARGET_SHELL | sudo tee -a $FTP_START_TARGET>/dev/null
+	echo 'Type=Application' | sudo tee -a $FTP_START_TARGET>/dev/null
+	echo 'Name=FTP_server' | sudo tee -a $FTP_START_TARGET>/dev/null
+	echo 'Terminal=true' | sudo tee -a $FTP_START_TARGET>/dev/null
+fi
